make P2PEditor application orchestrator
